Don't change laser.vector if reflected value undefined
This could happen if a new laser is added by clicking on a boundary cell. In that case the reflection detection will consider that the direction of the laser should be reflected, but in 50% of cases it's already pointing away from the boundary and doesn't need to be.

This fixes a bug whereby the lasers get 'stuck' as a single cell on the boundary in those cases.